+++ libparse/clk_sel240x.c
@@ -85,8 +85,8 @@ inp_sel240x( parse_t      *parseio,
 {
 	unsigned long rc;
 
-	parseprintf( DD_PARSE,
-	             ("inp_sel240x(0x%lx, 0x%x, ...)\n",(long)parseio, ch));
+	parseprintf(DD_PARSE, ("inp_sel240x(0x%lx, 0x%x, ...)\n",
+                    (unsigned long)parseio, (unsigned)ch));
 
 	switch( ch )
 	{
